+++ lib/formdata/index.ts
@@ -1,6 +1,8 @@
 import {isFormData} from './is';
+import {toObject} from './toObject';
 
 export {
     isFormData,
-    isFormData as is
+    isFormData as is,
+    toObject
 };