+++ test/lib/formdata/index.test.ts
@@ -2,12 +2,14 @@ import {describe, it}   from 'node:test';
 import * as assert      from 'node:assert/strict';
 import * as LibFormData from '../../../lib/formdata';
 import is               from '../../../lib/formdata/is';
+import toObject         from '../../../lib/formdata/toObject';
 
 describe('FormData - *', () => {
     it('Should be a correct export', () => {
         assert.deepEqual(LibFormData, {
             isFormData: is,
             is,
+            toObject,
         });
     });
 });